fix: move Qt5Port.h to GUI library, update fromQStringList callers
- Move Qt5Port.h from core (removed) to openms_gui/VISUAL/MISC/
- Add fromQStringList() to Qt5Port.h (replaces StringListUtils version)
- Update 3 GUI files and 2 test files to use new include path
- Remove DataValue(const QString&), toQString(), operator=(const QString&)
  test sections (APIs removed from DataValue)
