+++ src/tests/class_tests/openms/source/OPXLHelper_test.cpp
@@ -12,6 +12,7 @@
 ///////////////////////////
 
 #include <OpenMS/ANALYSIS/XLMS/OPXLHelper.h>
+#include <OpenMS/VISUAL/MISC/Qt5Port.h>
 
 #include <OpenMS/ANALYSIS/XLMS/OPXLDataStructs.h>
 #include <OpenMS/CHEMISTRY/TheoreticalSpectrumGeneratorXLMS.h>
@@ -46,8 +47,8 @@ QStringList q_str_list1;
 QStringList q_str_list2;
 q_str_list1 << "Carbamidomethyl (C)" << "Carbamidomethyl (T)";
 q_str_list2 << "Oxidation (M)" << "Oxidation (Y)";
-StringList fixedModNames = StringListUtils::fromQStringList(q_str_list1);
-StringList varModNames = StringListUtils::fromQStringList(q_str_list2);
+StringList fixedModNames = fromQStringList(q_str_list1);
+StringList varModNames = fromQStringList(q_str_list2);
 const ModifiedPeptideGenerator::MapToResidueType& fixed_modifications = ModifiedPeptideGenerator::getModifications(fixedModNames);
 const ModifiedPeptideGenerator::MapToResidueType& variable_modifications = ModifiedPeptideGenerator::getModifications(varModNames);
 
@@ -55,8 +56,8 @@ QStringList q_str_list3;
 QStringList q_str_list4;
 q_str_list3 << "K" << "E";
 q_str_list4 << "D" << "E" << "C-term";
-StringList cross_link_residue1 = StringListUtils::fromQStringList(q_str_list3);
-StringList cross_link_residue2 = StringListUtils::fromQStringList(q_str_list4);
+StringList cross_link_residue1 = fromQStringList(q_str_list3);
+StringList cross_link_residue2 = fromQStringList(q_str_list4);
 
 Size max_variable_mods_per_peptide = 5;
