+++ src/tests/class_tests/openms/source/StringListUtils_test.cpp
@@ -11,6 +11,7 @@
 ///////////////////////////
 
 #include <OpenMS/DATASTRUCTURES/StringListUtils.h>
+#include <OpenMS/VISUAL/MISC/Qt5Port.h>
 #include <QStringList>
 #include <OpenMS/DATASTRUCTURES/ListUtils.h>
 
@@ -28,7 +29,7 @@ START_SECTION((static StringList fromQStringList(const QStringList &rhs)))
   QStringList q_str_list;
   q_str_list << "First Element" << "Second Element" << "Third Element";
 
-  StringList str_list = StringListUtils::fromQStringList(q_str_list);
+  StringList str_list = fromQStringList(q_str_list);
   TEST_EQUAL((int)str_list.size(), q_str_list.size())
   ABORT_IF((int)str_list.size() != q_str_list.size())
   for(Size i = 0 ; i < str_list.size() ; ++i)
